feat: modify file structure
