--- Vesper-Agents-Suite/Maker-Strategy-Agent/src/maker.strategy.cases.agent.spec.ts
@@ -1,14 +1,0 @@
-import BigNumber from 'bignumber.js';
-import { Finding, HandleBlock, BlockEvent } from 'forta-agent';
-import provideMakerStrategyHandler from './maker.strategy.cases.agent';
-
-describe('Maker Strategy Agent Test Suit', () => {
-  const web3CallMock = jest.fn();
-  let handleBlock: HandleBlock;
-
-  beforeAll(() => {
-    handleBlock = provideMakerStrategyHandler(web3CallMock);
-  });
-
-  it('', () => {});
-});
